Make MessageTypeNotifier generic
